Refactor: move under agent functions to management package
This commit moves the UnderAgent, WithTraceLevel and other functions
from the publisher package to the management package.
