+++ libbeat/processors/urldecode/urldecode.go
@@ -23,10 +23,10 @@ import (
 	"net/url"
 
 	"github.com/elastic/beats/v7/libbeat/beat"
+	"github.com/elastic/beats/v7/libbeat/management"
 	"github.com/elastic/beats/v7/libbeat/processors"
 	"github.com/elastic/beats/v7/libbeat/processors/checks"
 	jsprocessor "github.com/elastic/beats/v7/libbeat/processors/script/javascript/module/processor"
-	"github.com/elastic/beats/v7/libbeat/publisher"
 	"github.com/elastic/elastic-agent-libs/config"
 	"github.com/elastic/elastic-agent-libs/logp"
 	"github.com/elastic/elastic-agent-libs/mapstr"
@@ -83,7 +83,7 @@ func (p *urlDecode) Run(event *beat.Event) (*beat.Event, error) {
 		err := p.decodeField(field.From, field.To, event)
 		if err != nil {
 			errMsg := fmt.Errorf("failed to decode fields in urldecode processor: %w", err)
-			if publisher.LogWithTrace() {
+			if management.TraceLevelEnabled() {
 				p.log.Debug(errMsg.Error())
 			}
 			if p.config.FailOnError {
